+++ includes/minishell.h
@@ -126,8 +126,6 @@ void			do_piping(int pipes[2], t_node *node, t_global_state *state);
 void			close_pipes(int pipes[2], t_node *node, t_global_state *state);
 void			close_parent_pipe(int pipes[2], t_node *node, t_global_state *state);
 
-void			init_redirect(t_redirect *redirect);
-int				is_redirect_token(t_token *token);
 int				set_redirect(t_list **tokens, t_redirect *rd, t_envs *envs);
 
 t_node			*preprocess(char *input, t_global_state *state, int debug);
@@ -159,6 +157,8 @@ int				is_command_token(t_list **token_list);
 t_node			*new_node(t_node *lhs, t_node *rhs, t_node_kind attr);
 int				consume_token(t_list **token_list, t_token_kind kind);
 
+int				execute_commands(t_node *node, int pipes[2], t_global_state *s);
+int				execute_pipe(t_node *ast, t_global_state *state);
 int				execute(t_node *ast, t_global_state *state);
 char			**construct_argv(t_list *tokens, t_global_state *state);
